+++ src/screens/AuthScreen/AuthScreen.tsx
@@ -6,10 +6,10 @@ import { StackNavigationOptions } from '@react-navigation/stack';
 import { Button } from 'react-native-elements';
 import { useDispatch } from 'react-redux';
 
-import PlainText from '../components/UI/PlainText';
-import { NavigationData } from '../navigation/AuthNavigator';
-import { loginWithFacebook, loginWithGoogle } from '../redux/auth/actions';
-import Colors from '../theme/Colors';
+import PlainText from '../../components/UI/PlainText';
+import { NavigationData } from '../../navigation/AuthNavigator';
+import { loginWithFacebook, loginWithGoogle } from '../../redux/auth/actions';
+import Colors from '../../theme/Colors';
 
 import {
   StyledButtonsContainer,
